fixed runtime mismatch /MTd vs /MDd vs /clr introduced vw_dynamic: same code as vw_static, but linked using MTd to be compatible with /clr added custom nuget target to deploy zlib.dll added documentation fixed c++/cli formatting

+++ vowpalwabbit/vw_exception.cc
@@ -52,7 +52,7 @@ void vw_trace(const char* filename, int linenumber, const char* fmt, ...)
 bool launchDebugger()
 {	// Get System directory, typically c:\windows\system32
 	std::wstring systemDir(MAX_PATH + 1, '\0');
-	UINT nChars = GetSystemDirectoryW(&systemDir[0], systemDir.length());
+	UINT nChars = GetSystemDirectoryW(&systemDir[0], (UINT)systemDir.length());
 	if (nChars == 0) return false; // failed to get system directory
 	systemDir.resize(nChars);
